+++ packages/asc-ui/src/components/Header/HeaderLogoText.tsx
@@ -3,7 +3,6 @@ import React, {
   FunctionComponent,
   HTMLAttributes,
 } from 'react'
-import { StyledProps } from 'styled-components'
 import HeaderLogoTextStyle from './HeaderLogoTextStyle'
 import HeaderTitle from './HeaderTitle'
 
@@ -14,15 +13,21 @@ export interface LogoProps extends AnchorHTMLAttributes<HTMLAnchorElement> {
 }
 
 interface HeaderLogoTextProps extends HTMLAttributes<HTMLHeadingElement> {
+  as?: string | React.ComponentType
   homeLink: string
   tall?: boolean
   logo?: FunctionComponent<LogoProps>
 }
 
-const HeaderLogoText: FunctionComponent<
-  HeaderLogoTextProps & StyledProps<any>
-> = ({ title, homeLink, tall, logo: LogoIcon, ...otherProps }) => (
-  <HeaderLogoTextStyle {...otherProps}>
+const HeaderLogoText: FunctionComponent<HeaderLogoTextProps> = ({
+  as,
+  title,
+  homeLink,
+  tall,
+  logo: LogoIcon,
+  ...otherProps
+}) => (
+  <HeaderLogoTextStyle as={as} {...otherProps}>
     {LogoIcon && <LogoIcon href={homeLink} title={title} tall={tall} />}
     {title && <HeaderTitle href={homeLink}>{title}</HeaderTitle>}
   </HeaderLogoTextStyle>