+++ packages/asc-ui/src/components/Select/index.ts
@@ -1,5 +1,6 @@
 export {
   default as SelectStyle,
   SelectWrapper as SelectWrapperStyle,
+  AbsoluteContentWrapper as SelectContentWrapperStyle,
 } from './SelectStyle'
 export { default } from './Select'