+++ packages/asc-ui/src/components/Header/Header.tsx
@@ -1,27 +1,33 @@
-import React from 'react'
+import React, {
+  ElementType,
+  FunctionComponent,
+  HTMLAttributes,
+  ReactNode,
+} from 'react'
+import { CustomCssPropsType } from '../../utils'
+import AmsterdamLogo from '../AmsterdamLogo'
+import HeaderLinks from './HeaderLinks'
+import HeaderLogoText, { LogoProps } from './HeaderLogoText'
+import HeaderNavigation from './HeaderNavigation'
 import HeaderStyle, { Props as HeaderStyleProps } from './HeaderStyle'
 import HeaderWrapperStyle, {
   Props as HeaderWrapperProps,
 } from './HeaderWrapperStyle'
-import HeaderLogoText, { LogoProps } from './HeaderLogoText'
-import HeaderNavigation from './HeaderNavigation'
-import HeaderLinks from './HeaderLinks'
-import { CustomCssPropsType } from '../../utils'
-import AmsterdamLogo from '../AmsterdamLogo'
 
-type Props = {
+type HeaderProps = {
   tall?: boolean
   homeLink: string
   title?: string
-  navigation?: React.ReactNode
-  links?: React.ReactNode
-  logo?: React.FC<LogoProps>
+  navigation?: ReactNode
+  links?: ReactNode
+  logo?: FunctionComponent<LogoProps>
+  headerLogoTextAs?: ElementType
 } & HeaderWrapperProps &
   HeaderStyleProps &
   CustomCssPropsType &
-  React.HTMLAttributes<HTMLDivElement>
+  HTMLAttributes<HTMLDivElement>
 
-const Header: React.FC<Props> = ({
+const Header: FunctionComponent<HeaderProps> = ({
   css,
   title,
   homeLink,
@@ -30,11 +36,15 @@ const Header: React.FC<Props> = ({
   navigation,
   links,
   logo,
+  headerLogoTextAs,
   ...otherProps
 }) => (
   <HeaderWrapperStyle {...{ css, tall }} id="header">
     <HeaderStyle {...{ fullWidth, ...otherProps }}>
-      <HeaderLogoText {...{ tall, title, homeLink, logo }} />
+      <HeaderLogoText
+        as={headerLogoTextAs}
+        {...{ tall, title, homeLink, logo }}
+      />
       <HeaderNavigation>{navigation}</HeaderNavigation>
       <HeaderLinks>{links}</HeaderLinks>
     </HeaderStyle>