+++ packages/asc-ui/src/components/Header/HeaderLogoText.tsx
@@ -1,28 +1,34 @@
-import React from 'react'
+import React, {
+  AnchorHTMLAttributes,
+  ElementType,
+  FunctionComponent,
+  HTMLAttributes,
+} from 'react'
 import HeaderLogoTextStyle from './HeaderLogoTextStyle'
 import HeaderTitle from './HeaderTitle'
 
-export interface LogoProps
-  extends React.AnchorHTMLAttributes<HTMLAnchorElement> {
+export interface LogoProps extends AnchorHTMLAttributes<HTMLAnchorElement> {
   href: string
   tall?: boolean
   title?: string
 }
 
-interface HeaderLogoTextProps extends React.HTMLAttributes<HTMLHeadingElement> {
+interface HeaderLogoTextProps extends HTMLAttributes<HTMLHeadingElement> {
+  as?: ElementType
   homeLink: string
   tall?: boolean
-  logo?: React.FC<LogoProps>
+  logo?: FunctionComponent<LogoProps>
 }
 
-const HeaderLogoText: React.FC<HeaderLogoTextProps> = ({
+const HeaderLogoText: FunctionComponent<HeaderLogoTextProps> = ({
+  as,
   title,
   homeLink,
   tall,
   logo: LogoIcon,
   ...otherProps
 }) => (
-  <HeaderLogoTextStyle {...otherProps}>
+  <HeaderLogoTextStyle as={as} {...otherProps}>
     {LogoIcon && <LogoIcon href={homeLink} title={title} tall={tall} />}
     {title && <HeaderTitle href={homeLink}>{title}</HeaderTitle>}
   </HeaderLogoTextStyle>