+++ packages/asc-ui/src/components/Toggle/Toggle.tsx
@@ -1,4 +1,15 @@
-import React, { HTMLAttributes } from 'react'
+import React, {
+  ButtonHTMLAttributes,
+  ElementType,
+  FunctionComponent,
+  KeyboardEvent,
+  HTMLAttributes,
+  MouseEvent,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from 'react'
 import ToggleStyle, { Props as ToggleStyleProps } from './ToggleStyle'
 import ownerDocument from '../../utils/ownerDocument'
 import usePassPropsToChildren from '../../utils/hooks/usePassPropsToChildren'
@@ -9,25 +20,25 @@ import ToggleButton, {
 import BackDrop, { Props as BackDropProps } from '../BackDrop/BackDrop'
 
 export type ToggleHandlerProps = {
-  as?: keyof JSX.IntrinsicElements | React.ComponentType<any>
+  as?: ElementType
 } & ToggleButtonProps &
-  React.HTMLAttributes<HTMLElement>
+  HTMLAttributes<HTMLElement>
 
 export type Props = {
   render?: boolean
   ariaLabel?: string
   onOpen?: (open: boolean) => void
-  ToggleHandler?: React.FC<
-    ToggleButtonProps & React.ButtonHTMLAttributes<HTMLButtonElement>
+  ToggleHandler?: FunctionComponent<
+    ToggleButtonProps & ButtonHTMLAttributes<HTMLButtonElement>
   >
   rotateOnOpen?: number
-  as?: keyof JSX.IntrinsicElements | React.ComponentType<any>
+  as?: ElementType
 } & ToggleStyleProps &
   ToggleHandlerProps &
   BackDropProps
 
 // Todo: refactor this to Collapse component https://github.com/Amsterdam/amsterdam-styled-components/issues/379
-const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
+const Toggle: FunctionComponent<Props & HTMLAttributes<HTMLDivElement>> = ({
   children: childrenProps,
   onClick,
   open: openProp,
@@ -45,9 +56,9 @@ const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
   ariaLabel,
   ...otherProps
 }) => {
-  const [open, setOpen] = React.useState(false)
+  const [open, setOpen] = useState(false)
 
-  const handleOnOpen = React.useCallback(
+  const handleOnOpen = useCallback(
     (openParam: boolean) => {
       setOpen(openParam)
 
@@ -63,7 +74,7 @@ const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
   const { onKeyDown: onKeyDownHook } = useActionOnEscape(() =>
     handleOnOpen(false),
   )
-  const ref = React.useRef<HTMLDivElement>(null!)
+  const ref = useRef<HTMLDivElement>(null!)
 
   const handleOnBlur = () => {
     setTimeout(() => {
@@ -77,14 +88,14 @@ const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
     })
   }
 
-  const handleOnKeyDown = (e: React.KeyboardEvent<HTMLElement>) => {
+  const handleOnKeyDown = (e: KeyboardEvent<HTMLElement>) => {
     onKeyDownHook(e)
     if (onKeyDown) {
       onKeyDown(e)
     }
   }
 
-  const handleOnClick = (e: React.MouseEvent<HTMLElement, MouseEvent>) => {
+  const handleOnClick = (e: MouseEvent<HTMLElement>) => {
     handleOnOpen(!open)
 
     if (onClick) {
@@ -97,7 +108,7 @@ const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
   })
 
   // Useful if parent needs to take over control the isOpen state
-  React.useEffect(() => {
+  useEffect(() => {
     if (typeof openProp !== 'undefined') {
       handleOnOpen(openProp)
     }