+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Prefix the change with one of these keywords:
 - _Fixed_: for any bug fixes.
 - _Security_: in case of vulnerabilities.
 
+## [UNRELEASED]
+
+- Added: `headerLogoTextAs` prop to `Header` component to overwrite the default `h1` tag
+
 ## [0.26.1]
 
 - Changed: Types for the `Alert` component (`AlertProps` and `AlertLevel`) are now exported from the package root.